With the removal of DDR as configuration requirement, bitstream has new
format.

This PR support generation of bitstream from JSON text to bitstream
binary. This is the most powerful flow that user can generate bitstream
in any way they want.

Next step will be 
1. bitstream generation from .bitasm (a.k.a SOF) 
2. Remove old bitstream format support from code base

+++ src/ConfigurationRS/CFGCrypto/CFGCrypto_key.cpp
@@ -391,7 +391,7 @@ size_t CFGCrypto_KEY::sign(const uint8_t* digest, const size_t digest_size,
   return signed_size;
 }
 
-uint32_t CFGCrypto_KEY::get_bitstream_signing_algo() {
+uint8_t CFGCrypto_KEY::get_bitstream_signing_algo() {
   CFG_ASSERT(m_key_info != nullptr);
   return m_key_info->bitstream_algo;
 }
